Add a very basic Lua console
Shift+` to access the console

The Lua interpreter will be used for scripting later

+++ src/util/io.h
@@ -14,7 +14,7 @@
 #include <SDL.h>
 #include <stdio.h>
 
-char* read_all(const char *filename, int *size);
+char* read_all(const char *filename, size_t *size);
 
 char* SDL_RWgets(SDL_RWops *rwops, char *buf, size_t bufsize);
 char* SDL_RWgets_realloc(SDL_RWops *rwops, char **buf, size_t *bufsize);
